requirements.txt
anyio==4.6.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
ollama==0.3.3
pymongo==4.8.0
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.2
Werkzeug==3.0.3
